refactor: BodyProfile, User 엔티티 개편 완료
