--- src/main/java/tukorea_2024_s3_10/eat_fit/domain/auth/Role.java
@@ -1,5 +1,0 @@
-package tukorea_2024_s3_10.eat_fit.domain.auth;
-
-public enum Role {
-    ROLE_GUEST, ROLE_USER
-}
