+++ src/main/java/tukorea_2024_s3_10/eat_fit/domain/user/entity/Role.java
@@ -1,0 +1,5 @@
+package tukorea_2024_s3_10.eat_fit.domain.user.entity;
+
+public enum Role {
+    ROLE_GUEST, ROLE_USER
+}
