+++ src/main/java/tukorea_2024_s3_10/eat_fit/infrastructure/user/Repository/DbUserProfileRepository.java
@@ -14,7 +14,7 @@ public class DbUserProfileRepository implements UserProfileRepository {
 
     @Override
     public Optional<BodyProfile> findById(Long id) {
-        return userProfileRepository.findById(id);
+        return null;
     }
 
     @Override
